+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/domain/factory/AuthorizationRequestFactory.java
@@ -4,6 +4,8 @@ import com.processing.merchantacquirer.domain.StanGenerator;
 import com.processing.merchantacquirer.domain.entity.Merchant;
 import com.processing.merchantacquirer.domain.entity.Terminal;
 import com.processing.merchantacquirer.domain.model.AuthorizationRequest;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -14,7 +16,7 @@ public class AuthorizationRequestFactory {
   private final StanGenerator stanGenerator;
 
   public AuthorizationRequest build(
-      String pan, Integer amount, Terminal terminal, Merchant merchant) {
+          String pan, BigDecimal amount, Terminal terminal, Merchant merchant) {
     LocalDateTime time = LocalDateTime.now();
 
     return AuthorizationRequest.builder()
